chore: 변경된 API 명세에 따라 기존 엔드포인트 변경 설정 및 테스트 코드 수정
