+++ src/main/kotlin/kr/hhplus/be/server/interfaces/balance/BalanceRequest.kt
@@ -1,5 +1,8 @@
 package kr.hhplus.be.server.interfaces.balance
 
 sealed class BalanceRequest {
-    data class Charge(val amount: Int)
+    data class Charge(
+        val customerId: Long,
+        val amount: Int
+    )
 }
